Change carry index styling, add modal for whispers
Next steps for modal whisper is to have it automatically copy the
whisper text to clipboard for easy copy pasta. Additionally, there needs
to be a validation on the user model to check if they have a main
character added and set as main before they can create a carry.