Upgrade to 6rc1, switch to RSpec, add char contact
The upgrade to rails 6.0.0.rc1 is in preperation for the upcomming rails
6 bump.

Now that the basic structure of the application is implemented I need to
start writing tests, which should've been written from the start.
Setting up RSpec is the first step.

Trying to cut down on the amount of queries needed for some pages. By
adding a character to carries I'm able to directly pull in the whisper
contact through that instead of a joins with users to get to the data. I
was already needing the character loaded for raider.io score so this cut
out a step.

Removed duplicate user queries on pages already authenticating the current
user.